+++ js/main.js
@@ -1,6 +1,5 @@
 import {renderScreen} from './utils';
 import {screenElement as welcomeScreenElement, initWelcomeScreen} from './screens/welcome';
-import {artistScreenElement, initArtistScreen, initGenreScreen} from './screens/game';
 // import {initFailTimeScreen, initFailTriesScreen, initSuccessScreen} from './screens/results';
