+++ js/screens/change-screen.js
@@ -1,10 +1,6 @@
-import {renderScreen} from "../utils";
-import {BUTTONS, getResultTemplate, RESULTS, TITLES as RESULTS_TITLES} from "./results";
 import {levels} from "../data/data";
-import {headerTemplate} from "../templates/header";
-import {getArtistTemplate} from "../templates/artist";
-import {getGenreTemplate} from "../templates/genre";
 import {artistScreen} from "../screens/artist-screen";
+import {genreScreen} from "./genre-screen";
 
 const MAX_QUESTIONS = 10;
 
@@ -17,10 +13,8 @@ export const changeScreen = (state) => {
     // renderScreen(getResultTemplate(RESULTS_TITLES.win, RESULTS.win, BUTTONS.win));
   } else if (levels[state.level].type === `artist`) {
     artistScreen(state);
-    // state.level++;
   } else if (levels[state.level].type === `genre`) {
-    // renderScreen(getGameTemplate(`genre`, headerTemplate, TITLES.genre.pop, getGenreTemplate(levels[state.level])));
-    // state.level++;
+    genreScreen(state);
   }
 };
